6-size.c

+++ 0x00-hello_world/6-size.c
@@ -15,6 +15,6 @@ printf("Size of a char: %zu byte(s)\n", (char) sizeof(a));
 printf("Size of an int: %zu byte(s)\n", (int)sizeof(b));
 printf("Size of a long int: %zu byte(s)\n", (long int)sizeof(c));
 printf("Size of a long long: %zu byte(s)\n", (long long)sizeof(d));
-printf("Size of a float: %zu byte(s)\n", (float)sizeof(f));
+printf("Size of a float: %d byte(s)\n", sizeof(f));
 return (0);
 }
